+++ Pins b2.02.txt
@@ -1,12 +1,23 @@
 ﻿Pins - Funktion RJ45 (Encoder A,A' / B, B' / Z,Z')
-8- GND 	BR
-7- 5V	BR/W
-6- B+	GR
-5- Z-	BL/w
-4- Z+	BL
-3- B-	GR/W
-2- A- 	OR
-1- A+	OR/W
+8- GND
+7- 5V
+6- B+
+5- Z-
+4- Z+
+3- B-
+2- A-
+1- A+
+---------------
+
+Pins - Funktion RJ45 ( 2A,A' / 2B, B')
+8- GND
+7- 5V
+6- B+
+5- Switch Arduino 6 
+4- Switch Arduino 7 
+3- B-
+2- A-
+1- A+
 ---------------
 Pins - Funktion RJ45 (Encoder A/B/Z und Eingänge (Endschalter) )
 8- GND
@@ -18,15 +29,26 @@ Pins - Funktion RJ45 (Encoder A/B/Z und Eingänge (Endschalter) )
 2- Arduino 38
 1- A+
 --------------
-Pins - Funktion RJ45 (Endstufe)
+Pins - Funktion RJ45 MotorA
 8- GND
-7- Arduino 11 (Status Eingang)
+7- Arduino 4 (Status Eingang)
 6- 5V
-5- Arduino 8 Dir Pin
+5- Arduino 8 (Ausganag Enable)
 4- 5V
-3- Arduino 9 Step Pin
+3- Arduino 9 Dir Pin
 2- GND
-1- Arduino 10 Enable
+1- Arduino 10 Step Pin
+
+--------------
+Pins - Funktion RJ45 MotorB
+8- GND
+7- Arduino 5 (Status Eingang)
+6- 5V
+5- Arduino 11 (Ausganag Enable)
+4- 5V
+3- Arduino 12 Dir Pin
+2- GND
+1- Arduino 13 Step Pin
 
 ----------------
 S1- Arduino 42
@@ -42,11 +64,15 @@ S10- Adruino 51
 S11- Arduino 40 (Encoder Taster)
 S12 - Reset
 
-Encoder A- Arduino 19 Int4
-Encoder B- Arduino 20 Int3
+Encoder A- Arduino 18 Int5
+Encoder B- Arduino 17 
 
 Externer Encoder A- Arduino 2 Int0
 Externer Encoder B- Arduino 3 Int1
+Externer Encoder Z- Arduino 21 Int2
+
+Externer Encoder2 A- Arduino 19 Int3
+Externer Encoder2 B- Arduino 20 Int4
 
 Endschalter - 38
 Endschalter - 39
@@ -57,3 +83,4 @@ LED Status 2- Arduino 52
 LED Status 3- Arduino 30
 LED Status 4- Arduino 31
 Summer- Arduino 32
+Relay - Arduin A7
