--- Version b0.1/README.md
@@ -1,6 +1,0 @@
-ZyklenAutomatik
-==========
-Elektronisches Getriebe mit Zyklen Funktion auf auf Arduino Mega2560 Basis. Zur Nachbildung eines Getriebes für die Leitspindel einer Drehbank
-
-
-<a rel="license" href="http://creativecommons.org/licenses/by-nc-sa/3.0/deed.de"><img alt="Creative Commons Lizenzvertrag" style="border-width:0" src="http://i.creativecommons.org/l/by-nc-sa/3.0/88x31.png" /></a><br />Dieses Werk bzw. Inhalt ist lizenziert unter einer <a rel="license" href="http://creativecommons.org/licenses/by-nc-sa/3.0/deed.de">Creative Commons Namensnennung - Nicht-kommerziell - Weitergabe unter gleichen Bedingungen 3.0 Unported Lizenz</a>.
