+++ Zyklenautomatik.atsln
@@ -1,7 +1,7 @@
 ﻿
 Microsoft Visual Studio Solution File, Format Version 11.00
 # Atmel Studio Solution File, Format Version 11.00
-Project("{E66E83B9-2572-4076-B26E-6BE79FF3018A}") = "Bresenham_Shield_muck", "Bresenham_Shield_muck.cppproj", "{07B459B4-8D76-4638-BC6D-A459BA2D616B}"
+Project("{E66E83B9-2572-4076-B26E-6BE79FF3018A}") = "Zyklenautomatik", "Zyklenautomatik.cppproj", "{07B459B4-8D76-4638-BC6D-A459BA2D616B}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
